feat(chat)!: redesign the conversation transfer toast around files and real progress
